Add fuse_marlin_moe test to ci and add new ep test (#14686)
